src/requirements.txt
azure-functions
azure-storage-blob
azure-storage-queue
visionone-filesecurity
requests
